+++ extensions/spotify-player/src/components/TrackActionPanel.tsx
@@ -1,8 +1,6 @@
 import { Action, ActionPanel, Icon } from "@raycast/api";
 import { SimplifiedAlbumObject, SimplifiedTrackObject } from "../helpers/spotify.api";
 import { TracksList } from "./TracksList";
-import { useMyPlaylists } from "../hooks/useMyPlaylists";
-import { useMe } from "../hooks/useMe";
 import { AddToPlaylistAction } from "./AddToPlaylistAction";
 import { FooterAction } from "./FooterAction";
 import { AddToQueueAction } from "./AddtoQueueAction";
@@ -30,9 +28,6 @@ export function TrackActionPanel({
   playingContext,
   tracksToQueue,
 }: TrackActionPanelProps) {
-  const { myPlaylistsData } = useMyPlaylists();
-  const { meData } = useMe();
-
   return (
     <ActionPanel>
       <PlayAction id={track.id as string} type="track" playingContext={playingContext} tracksToQueue={tracksToQueue} />
@@ -47,9 +42,7 @@ export function TrackActionPanel({
       <StartRadioAction trackId={track.id} />
       {showAddToSaved && <AddToSavedTracksAction trackId={track.id} />}
       {track.uri && <AddToQueueAction uri={track.uri} title={title} />}
-      {myPlaylistsData?.items && meData && track.uri && (
-        <AddToPlaylistAction playlists={myPlaylistsData.items} meData={meData} uri={track.uri} />
-      )}
+      {track.uri && <AddToPlaylistAction uri={track.uri} />}
       <FooterAction url={track?.external_urls?.spotify} uri={track.uri} title={title} />
     </ActionPanel>
   );